Implement findEnclosure algorithm.
This finds a region of tiles enclosed by 'walls' using a flood fill approach.
When a more general flood fill function is implemented, this should be based on
that algorithm.

The general flood fill should return a range that can be stepped through the
process one tile at a time.